REFACTO : moved the collection item models into core/services
